+++ python2verilog/simulation/__init__.py
@@ -2,4 +2,4 @@
 Wrappers for simulation tools
 """
 
-from .display_parse import UnknownValue, parse_stdout, strip_signals
+from .display import UnknownValue, parse_stdout, strip_signals
