+++ tests/simulation/test_simulation.py
@@ -1,3 +1,4 @@
+import logging
 import unittest
 from pathlib import Path
 
@@ -5,8 +6,7 @@ from python2verilog.api import verilogify
 from python2verilog.api.modes import Modes
 from python2verilog.api.namespace import namespace_to_verilog, new_namespace
 from python2verilog.api.verilogify import get_context
-from python2verilog.simulation import iverilog
-from python2verilog.simulation.display_parse import parse_stdout, strip_signals
+from python2verilog.simulation import iverilog, parse_stdout, strip_signals
 from python2verilog.utils.fifo import temp_fifo
 
 goal_namespace = new_namespace(Path(__file__).parent / "dup_range_goal")
@@ -42,14 +42,14 @@ class TestSimulation(unittest.TestCase):
 
         module, testbench = namespace_to_verilog(goal_namespace)
         with temp_fifo() as module_fifo, temp_fifo() as tb_fifo:
-            cmd = iverilog.make_cmd("dup_range_goal_tb", [module_fifo, tb_fifo])
-            print(cmd)
-            stdout, err = iverilog.run_cmd_with_fifos(
-                cmd, {module_fifo: module, tb_fifo: testbench}, timeout=3
+            stdout, err = iverilog.run_with_fifos(
+                "dup_range_goal_tb",
+                {module_fifo: module, tb_fifo: testbench},
+                timeout=3,
             )
             self.assertFalse(err)
+            logging.warning(stdout)
             actual = list(strip_signals(parse_stdout(stdout)))
-            print(actual)
             expected = []
             test_cases = list(get_context(dup_range_goal).test_cases)
             for test in test_cases:
